+++ tests/Support/TestModels/TestModel.php
@@ -4,12 +4,12 @@ namespace Spatie\Medialibrary\Tests\Support\TestModels;
 
 use Illuminate\Database\Eloquent\Model;
 use Spatie\Medialibrary\HasMedia\HasMedia;
-use Spatie\Medialibrary\HasMedia\HasMediaTrait;
+use Spatie\Medialibrary\HasMedia\InteractsWithMedia;
 use Spatie\Medialibrary\Models\Media;
 
 class TestModel extends Model implements HasMedia
 {
-    use HasMediaTrait;
+    use InteractsWithMedia;
 
     protected $table = 'test_models';